Address latest comments
